feat(nvim-floaterm): add plugin
I was missing the ability to toggle a terminal quickly. During
development it happens often that I run a command and need to close and
reopen the terminal later. With the current tools provided by lazyvim
it is not possible to toggle the terminal.

I would have preferred to use toggleterm.nvim but it does not have
native support for fixing nested nvim sessions when using tools like
lazygit or ranger. Instead it is outsourcing this to "nvr".