fix(nvim-clipboard): use wayland copy paste for WSL
Coping with win32yank is somehow not working anymore. Neovim suggests in
':help clipboard-wsl' to use 'clip.exe' and a 'powershell' command but
this is causing problems in the yanky [2] plugin. Another solution [1]
suggests to use wl-clipboard which works fine but the tool must be
installed. Also this solution only works on newer WSL versions with WSLg
enabled.

[1]: https://stackoverflow.com/a/76388417
[2]: https://www.lazyvim.org/extras/coding/yanky